Change all variables except colour to floats

+++ pool/abstractbuilder.h
@@ -8,9 +8,10 @@
 class AbstractBuilder
 {
 public:
-    virtual Ball* buildBall(Coordinate coordinate, std::string color, int mass, int radius, double xVelocity, double yVelocity) = 0;
-    virtual Table* buildTable(unsigned int length, unsigned int height, unsigned int thickness, unsigned int space, unsigned int friction) = 0;
+    virtual Ball* buildBall(Coordinate coordinate, std::string color, float mass, float radius, float xVelocity, float yVelocity) = 0;
+    virtual Table* buildTable(float length, float height, float thickness, float space, float friction) = 0;
     virtual Pool getPool() = 0;
 };
 
 #endif // ABSTRACTBUILDER_H
+
